Moving AddMember to components and making placeholder editable
