+++ src/routes/App/routes/bill/routes/debtors/components/DebtorsPage.js
@@ -5,7 +5,7 @@ import { SubTitle, Title } from "components/text/index";
 import { Page } from "components/Container";
 import { List, ListDivider, ListItem } from "react-toolbox/lib/list/index";
 import { connect } from "react-redux";
-import AddMember from "../../../../new/components/AddMember/AddMember";
+import AddMember from "components/AddMember";
 import Avatar from "../../../../../../../components/Avatar/Avatar";
 import LocationBuilder from "../../../../../modules/navigation/LocationBuilder";
 import withNavigation from "../../../../../../../components/hoc/withNavigation";
