+++ src/routes/App/routes/new/components/AddBillForm.js
@@ -8,7 +8,7 @@ import inputTheme from "./InputTheme.scss";
 import headerButtonTheme from "./HeaderButtonTheme.scss";
 import { CustomShareListItem, EqualShareListItem, PercentalShareListItem } from "./ShareList";
 import HOList from "containers/HOList";
-import AddMember from "./AddMember";
+import AddMember from "components/AddMember";
 import AddBillProgressBarTheme from "./AddBillProgressBarTheme.scss"
 import { Link } from "react-router";
 import LocationBuilder from "../../../modules/navigation/LocationBuilder";
